refactor: 출력 방식 수정
- AbstractOperation를 연산 구조체들의 calculate 메서드가 연산값을 출력
- Operator 열거형의 원시값 삭제
- ReminderOperation 나머지 연산 구조체 삭제
- DivideOperation 구조체의 calculate 메서드에서 secondNumber(나누는 수)가 0일 경우 "0으로는 나눌 수 없습니다." 출력
- 나눗셈 테스트 코드 추가
